requirements.txt
requests
python-dotenv
yt-dlp
ffmpeg-python
openai-whisper